added cleanupkeys action and tests
